unstable: major changes and refactor

+++ .plok/install.sh
@@ -1,5 +1,8 @@
 main () {
-    rm pnpm-lock.yaml
+    if [ -f "pnpm-lock.yaml" ]; then
+        rm pnpm-lock.yaml
+    fi
+
     pnpm i
     bash .plok/log.sh install node_modules
 }
